+++ framy_generator/test/generators/framy_app_generator_test.dart
@@ -1,3 +1,4 @@
+import 'package:framy_generator/config/framy_config.dart';
 import 'package:framy_generator/framy_object.dart';
 import 'package:framy_generator/generator/framy_app_generator.dart';
 import 'package:test/test.dart';
@@ -5,27 +6,27 @@ import 'package:test/test.dart';
 void main() {
   group('FramyApp Generator result', () {
     test('should contain FramyApp class', () {
-      final result = generateFramyApp([], false);
+      final result = generateFramyApp([], false, FramyConfig());
       expect(result.contains('class FramyApp'), isTrue);
     });
 
     test('should contain MaterialApp', () {
-      final result = generateFramyApp([], false);
+      final result = generateFramyApp([], false, FramyConfig());
       expect(result.contains('MaterialApp'), isTrue);
     });
 
     test('should contain FramyApp Key', () {
-      final result = generateFramyApp([], false);
+      final result = generateFramyApp([], false, FramyConfig());
       expect(result.contains('Key(\'FramyApp\')'), isTrue);
     });
 
     test('should contain onGenerateRoute', () {
-      final result = generateFramyApp([], false);
+      final result = generateFramyApp([], false, FramyConfig());
       expect(result.contains('onGenerateRoute'), isTrue);
     });
 
     test('should not contain theme when empty list passed', () {
-      final result = generateFramyApp([], false);
+      final result = generateFramyApp([], false, FramyConfig());
       expect(result.contains('theme'), isFalse);
     });
 
@@ -37,7 +38,7 @@ void main() {
         ..kind = ElementKind.FUNCTION
         ..isStatic = true;
       //when
-      final result = generateFramyApp([themeDataObject], false);
+      final result = generateFramyApp([themeDataObject], false, FramyConfig());
       //then
       expect(result.contains('theme: getThemeData(),'), isTrue);
     });
@@ -51,7 +52,7 @@ void main() {
         ..kind = ElementKind.METHOD
         ..parentObject = (FramyObject()..name = 'CustomAppTheme');
       //when
-      final result = generateFramyApp([themeDataObject], false);
+      final result = generateFramyApp([themeDataObject], false, FramyConfig());
       //then
       expect(
           result.contains('theme: CustomAppTheme().getThemeData(),'), isTrue);
@@ -67,7 +68,7 @@ void main() {
         ..kind = ElementKind.METHOD
         ..parentObject = (FramyObject()..name = 'CustomAppTheme');
       //when
-      final result = generateFramyApp([themeDataObject], false);
+      final result = generateFramyApp([themeDataObject], false, FramyConfig());
       //then
       expect(result.contains('theme: CustomAppTheme.getThemeData(),'), isTrue);
     });
@@ -81,7 +82,7 @@ void main() {
         ..kind = ElementKind.GETTER
         ..parentObject = (FramyObject()..name = 'CustomAppTheme');
       //when
-      final result = generateFramyApp([themeDataObject], false);
+      final result = generateFramyApp([themeDataObject], false, FramyConfig());
       //then
       expect(result.contains('theme: CustomAppTheme().theme,'), isTrue);
     });
@@ -102,14 +103,15 @@ void main() {
         ..kind = ElementKind.GETTER
         ..parentObject = parentObject;
       //when
-      final result = generateFramyApp([colorObject, themeObject], false);
+      final result =
+          generateFramyApp([colorObject, themeObject], false, FramyConfig());
       //then
       expect(result.contains('theme: AppThemeWithGetter.themeData,'), isTrue);
     });
 
     test('should contain declaration of framyAppStateKey', () {
       //when
-      final result = generateFramyApp([], false);
+      final result = generateFramyApp([], false, FramyConfig());
       //then
       expect(
           result.contains(
@@ -119,16 +121,30 @@ void main() {
 
     test('should contain class FramyAppSettings', () {
       //when
-      final result = generateFramyApp([], false);
+      final result = generateFramyApp([], false, FramyConfig());
       //then
       expect(result.contains('class FramyAppSettings'), isTrue);
     });
 
     test('should add a ProviderScope when there are riverpod dependencies', () {
       //when
-      final result = generateFramyApp([], true);
+      final result = generateFramyApp([], true, FramyConfig());
       //then
       expect(result.contains('ProviderScope'), isTrue);
+    });
+
+    test('should use default values from FramyConfig', () {
+      final result = generateFramyApp(
+        [],
+        false,
+        FramyConfig()
+          ..wrapWithCenter = true
+          ..wrapWithScaffold = false
+          ..wrapWithSafeArea = null,
+      );
+      expect(result.contains('bool _wrapWithCenter = true;'), isTrue);
+      expect(result.contains('bool _wrapWithScaffold = false;'), isTrue);
+      expect(result.contains('bool _wrapWithSafeArea = null;'), isTrue);
     });
   });
 }