+++ framy_generator/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.2.2]
+
+* Added option to configure default values in `framy.yaml` or `pubspec.yaml` [#13](https://github.com/Fidev-io/framy/issues/13)
+
 ## [0.2.1]
 
 * Added option to resize dependencies panel [#70](https://github.com/Fidev-io/framy/issues/70)